fix(cache): one reading of CACHE_TTL_HOURS, and a typo no longer 500s
The variable was already wired, but six times over: five copies of
`int(os.getenv("CACHE_TTL_HOURS", 24)) * 3600` inline at the `cache.set` calls
and a sixth in `/api/config`. Six readings of one setting are six chances for
the TTL the settings panel prints to stop being the TTL the cache was given, and
the sixth had a sharper edge — `int("24h")` raises on a request path, so a stray
character in one environment variable presented as an app that would not load.

Now `cache_ttl_hours()`, read per call so a restart is enough, with every
`cache.set` and the config endpoint going through it. Anything unusable falls
back to 24 and says why, rather than raising: the worst honest consequence of
this setting is fetching a title again earlier than necessary.

Two behaviours it gains, both of them the point of being able to set it at all:

- Fractions. `0.5` is thirty minutes; `int()` could not express it.
- `0` means reuse nothing, which is what you want while chasing a wrong match.
  Pinned against diskcache itself, since `expire=0` could as easily have meant
  "no expiry" — which is what `expire=None` means.

Two things deliberately do not follow it. A "nothing matched" answer stays
capped at an hour, following the setting down but never up, because it is the
answer most likely to be wrong for a reason outside this app. And the TVDB auth
token keeps its own 24h: shortening the TTL to re-check a title should not mean
re-authenticating on every scan.

`ConfigOut.cache_ttl_hours` becomes a float, or the panel would round a
half-hour to a number the cache is not using.

+++ backend/models.py
@@ -80,7 +80,9 @@ class ConfigOut(BaseModel):
     media_roots: list[str]
     default_directory: str | None
     language_preference: list[str]
-    cache_ttl_hours: int
+    # Float, because `CACHE_TTL_HOURS=0.5` is a legitimate half-hour and rounding it
+    # to an int here would make the panel report a TTL the cache is not using.
+    cache_ttl_hours: float
     cache_entries: int
     cache_size_bytes: int
     thresholds: ThresholdsOut
